backend/pyproject.toml
[project]
name = "wellbee-backend"
version = "1.0.0"
description = "WellBee — Employee Wellness & Engagement Platform API"
requires-python = ">=3.12, <3.14"
dependencies = [
    "beanie==2.0.1",
    "fastapi==0.128.0",
    "google-genai==1.56.0",
    "motor==3.7.1",
    "passlib[bcrypt]==1.7.4",
    "protobuf==6.33.2",
    "pydantic==2.12.5",
    "pydantic-settings==2.7.1",
    "PyJWT==2.10.1",
    "python-dotenv==1.2.1",
    "slowapi==0.1.9",
    "tenacity==9.0.0",
    "uvicorn[standard]==0.24.0",
]

[dependency-groups]
dev = ["ruff>=0.8.0"]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
